Add a higher order function that performs repository standard calls
This higher‑order function accepts a callback [onConnect] that performs the repository call (returning a [RepositoryFutureResponse<T>]). It first checks for an internet connection; if there isn’t one, [onDisconnect] will perform a call if it's not null or it will
returns a [NoInternetConnectionFailure].
Otherwise, it executes the [onConnect] inside a try/catch block. If an exception is caught, an [UnknownFailure] is returned.